optimize
for some reason. the motor driver burn frequently, so i decide to delete
some unused code to see whether the situation improve or not.

for the level 1, reduce the test duty, since in the testing mode, the
duty too high seems not good idea.

for level 6, remove the i2c and dubounce code.
and add line to judge the auto pwm width, so that it can control whether
the pwm should send to he mosfet driver or not.